buyv_admin/requirements.txt
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Babel==4.0.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==21.2.0
WTForms==3.1.1
email-validator==2.1.0
